Hook sanitizers into the benchmark testing system.

I was slightly mistaken in the earlier sanitizer pull request, in that setting up unit/integration tests for different "mode" builds are more complicated. This pull request sets things up so the test system can run the sanitizer unit/integration tests. (They currently fail, but they run.) 

Also, we've been rather overzealous in the integration test ignore settings. e.g. we excluded "time", which will also exclude lines containing "runtime error" or other such ("antimetabolite", for example). I tweaked the ignore exceptions to be more specific, as a consequence massive integration test changes are expected, as lines which were ignored will be ignored no longer.

If there is some tracer line which I revealed which should probably be hidden, feel free to add a specific line to the ignore_list: just be as specific with the regex as possible, to avoid inadvertent hiding of lines. 

Old repository SHA1: f279414251c3e9759f756fa5e9af3ce03c4d3570

+++ source/ninja_build.py
@@ -22,7 +22,7 @@ class MyParser(argparse.ArgumentParser):
 def clean_build_path(build_path):
     for root, dirs, files in os.walk(build_path, topdown=False):
         for name in files:
-            if name != "CMakeLists.txt":
+            if name != "CMakeLists.txt" and name != "buildme.sh":
                 os.remove(os.path.join(root, name))
         for name in dirs:
             os.rmdir(os.path.join(root, name))
